Created Location class.

+++ lib/types/constants.dart
@@ -1,8 +1,8 @@
 class Constants {
   // Game.
-  static const double gridStart = -3.0;
-  static const double gridEnd = 3.0;
-  static const double gridSize = gridEnd - gridStart + 1;
+  static const int gridStart = -3;
+  static const int gridEnd = 3;
+  static const int gridSize = gridEnd - gridStart + 1;
   static const double cellOffset = 2.0;
   static const int matchesRequired = 3;
